Link stylesheet, nest/layout sidebar & header
In this commit, I fixed the error caused by neglecting to link the
stylesheet in the HTML head. Then, I nested child elements in the
sidebar and header sections of the page, and laid them out using mostly
grid, with a bit of flexbox within each item in the sidebar.