Nest/layout main-content
In this commit, I nexted child elements in the main-content sections, as
well as child elements within those. This fleshes out the rest of the
content for the page, including project cards and windows for shoutouts
and announcements. Similar to the content in the header and sidebar
containers, most of the content is laid out using grid, with a bit of
flexbox here and there for 1-dimensional layouts that only pertain to
alignment.