Add assets, begin theme styling
In this commit, I have added various assets, which will be credited in
the README file, including DM Sans (font-family hosted by Google Fonts),
a color theme with WCAG accessible contrast ratio, icons from
Pictogrammers, and a hatching dinosaur icon in a similar iconography
style from Marz Gallery on Freepik, which I will use as the logo for the
site. I have removed the borders since applying background-color to the
appropriate sections. I have also corrected a few layout issues,
replaced the placeholders with the appropriate icons, and ensured the
correct sizing for all of the icons.