Styling (misc)
In this commit, I have applied miscellaneous styling properties to fix
noted issues throughout the page. The sidebar, cards, and header have
had sizing and fit-content properties applied to fix layout issues. In
addition, the buttons and search bar have also had style and sizing
properties applied. I changed the cursor to be a pointer on the icons,
so that these appear usable in the same way as buttons to the user.
Finally, a few declarations within the stylesheet have been consolidated
together based on shared properties, for readability.